requirements.txt
python-telegram-bot==21.6
aiohttp==3.9.5
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
